Rename EWasmAST to WasmAST
